Update `SecretField` and `BlockSecret`
