+++ rnd/autogpt_server/autogpt_server/blocks/reddit.py
@@ -2,19 +2,19 @@ from datetime import datetime, timedelta, timezone
 
 import praw
 from typing import Any
-from pydantic import BaseModel, Field, SecretStr
+from pydantic import BaseModel, Field
 from typing import Iterator
 
 from autogpt_server.data.block import Block, BlockOutput, BlockSchema
 from autogpt_server.util.mock import MockObject
-from autogpt_server.data.config import SecretField
+from autogpt_server.data.model import BlockSecret, SecretField
 
 
 class RedditCredentials(BaseModel):
-    client_id: SecretStr = SecretField(key="reddit_client_id")
-    client_secret: SecretStr = SecretField(key="reddit_client_secret")
-    username: SecretStr = SecretField(key="reddit_username")
-    password: SecretStr = SecretField(key="reddit_password")
+    client_id: BlockSecret = SecretField(key="reddit_client_id")
+    client_secret: BlockSecret = SecretField(key="reddit_client_secret")
+    username: BlockSecret = SecretField(key="reddit_username")
+    password: BlockSecret = SecretField(key="reddit_password")
     user_agent: str | None = None
