+++ rnd/autogpt_server/autogpt_server/data/block.py
@@ -1,74 +1,16 @@
 from abc import ABC, abstractmethod
-from typing import Any, ClassVar, Generator, Generic, Optional, Type, TypeVar, cast
+from typing import Any, ClassVar, Generator, Generic, Type, TypeVar, cast
 
-from autogpt_server.util.settings import Secrets
 import jsonref
 import jsonschema
 from prisma.models import AgentBlock
-from pydantic import BaseModel, GetCoreSchemaHandler
-from pydantic_core import CoreSchema, core_schema
+from pydantic import BaseModel
 
 from autogpt_server.util import json
 
 BlockInput = dict[str, Any]
 BlockData = tuple[str, Any]
 BlockOutput = Generator[BlockData, None, None]
-
-
-class BlockFieldSecret:
-    def __init__(self, key=None, value=None):
-        if value is not None:
-            self._value = value
-            return
-
-        self._value = self.__get_secret(key)
-        if self._value is None:
-            raise ValueError(f"Secret {key} not found.")
-
-    STR: ClassVar[str] = "<secret>"
-    SECRETS: ClassVar[Secrets] = Secrets()
-
-    def __repr__(self):
-        return BlockFieldSecret.STR
-
-    def __str__(self):
-        return BlockFieldSecret.STR
-
-    @staticmethod
-    def __get_secret(key: str | None):
-        if not key or not hasattr(BlockFieldSecret.SECRETS, key):
-            return None
-        return getattr(BlockFieldSecret.SECRETS, key)
-
-    def get(self):
-        return str(self._value)
-
-    @classmethod
-    def parse_value(cls, value: Any) -> "BlockFieldSecret":
-        if isinstance(value, BlockFieldSecret):
-            return value
-        return BlockFieldSecret(value=value)
-    
-    @classmethod
-    def __get_pydantic_json_schema__(
-            cls, source_type: Any, handler: GetCoreSchemaHandler) -> dict[str, Any]:
-        return {
-            "type": "string",
-            "title": "BlockFieldSecret",
-            "description": "A secret field",
-        }
-
-    @classmethod
-    def __get_pydantic_core_schema__(
-            cls, source_type: Any, handler: GetCoreSchemaHandler) -> CoreSchema:
-        validate_fun = core_schema.no_info_plain_validator_function(cls.parse_value)
-        return core_schema.json_or_python_schema(
-            json_schema=validate_fun,
-            python_schema=validate_fun,
-            serialization=core_schema.plain_serializer_function_ser_schema(
-                lambda val: BlockFieldSecret.STR
-            ),
-        )
 
 
 class BlockSchema(BaseModel):
